Version 15/01/15
avec autoload et fonction dynamic à changer dans controller
ne fonctionne pas, la classe (VisiteurController par ex) n'est pas
trouvée

+++ src/Autoload.php
@@ -1,27 +1,28 @@
 <?php
 
-/*
+
 class Autoload
 {
     public static $nb = 0; // permet de compter le nbre de fois où l'on passe dans l'autoload 
     public static function className($className)
     {
+       if(class_exists($className)){
         echo '<pre>' . self::$nb . ' - Autoload : ' . $className;
         $check = explode('\\', $className); // me permet de créer un tableau Array de ce qui est passé en paramètres ($className)
         var_dump($check);
-        if($check[0] == 'backoffice')
-            $path = __DIR__ . '\..\backoffice\\' . $className . '.php'; // si la classe contient BackOffice, je vais faire un require sur le dossier src
-        else
-            $path = __DIR__ . '\\' . $className . '.php'; // sinon je vais faire un require direct (donc dans le dossier Vendor)
+        //à corriger -> on va tout le temps dans le dossier controller\
+           $path = __DIR__ . '\\'  . $className . '.php'; // sinon je vais faire un require direct (donc dans le dossier Vendor)
         
         echo "\n => $path</pre><hr>";
         
         require $path;        
         self::$nb++;
+        } else {
+           echo 
+            "<br>La classe n'existe pas."; 
+        }
+        
     }
 }
 spl_autoload_register(array('Autoload', 'className')); 
 // $u = new Employe\Nom_De_LaClasse;
- * 
- * 
- */
